refactor and move code related to generating piece moves yaml.  also change yaml representation for special moves (summon and divination).
