+++ yamlpieces.yaml
@@ -78,8 +78,7 @@ Duchess:
     - {x: 0, y: 1}
     - {x: 0, y: -1}
     - {x: -1, y: 0}
-    summon:
-    - {x: 0, y: 222}
+    special: summon
   front:
     command_square:
     - {x: 0, y: 1}
@@ -87,8 +86,7 @@ Duchess:
     - {x: 0, y: 1}
     - {x: 0, y: -1}
     - {x: -1, y: 0}
-    summon:
-    - {x: 0, y: 222}
+    special: summon
 Duke:
   back:
     slide:
@@ -189,9 +187,7 @@ Marshall:
     - {x: 0, y: 1}
     - {x: 0, y: -1}
 Oracle:
-  back:
-    divination:
-    - {x: 0, y: 222}
+  back: {special: divination}
   front:
     move:
     - {x: 1, y: 1}
